Cleared github files from library and created singleton for Http security handling
